accounts: move accounts to own module
